+++ Changelog.md
@@ -1,2 +1,3 @@
 ## 30.05.2024
 1. Forked from go-template
+2. Added forta-webhook support